backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==7.5.3
ecdsa==0.19.1
fastapi==0.115.11
fastmcp==0.4.1
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.48
langchain-google-genai==2.1.1
langchain-mcp-adapters==0.0.5
langchain-nvidia-ai-endpoints==0.3.9
langchain-openai==0.3.10
langchain-postgres==0.0.13
langchain-text-splitters==0.3.7
langgraph==0.3.20
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.6
langgraph-sdk==0.1.58
langgraph-supervisor==0.0.12
langsmith==0.3.18
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.4.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.68.2
orjson==3.10.15
ormsgpack==1.9.0
packaging==24.2
pgvector==0.3.6
pillow==11.1.0
primp==0.14.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.6
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
